Fix MergeFrontier.remove_failure().
MergeFrontier.remove_failure() was not doing the right thing.  If
there is a conflict like this:

    *******
    *......
    *.....X
    *......

then it should attempt to split the conflicting block into two blocks
(one of which might be empty) like this:

    *****||
    -----++
    *....|X
    -----+.

but instead it is doing this:

    *****|*
    *----+.
    *.....X
    *......

which is nonsense.  This could lead to an infinite loop (GitHub issue