Remove unused diagram argument, and s/fill_diagram/create_diagram/
The diagram argument to the fill_diagram() methods of MergeFrontier and
Block was never used, so remove it. Once removed, the fill_diagram()
name no longer makes much sense, since the diagram is created within
the method, and no longer passed in as an argument. Therefore, rename
the methods to create_diagram().

The docstrings of the two methods are adjusted accordingly.