Extend log message when simplifying to rebase-with-history.
When simplifying to rebase-with-history, append a note to the log
message explaining where the commit came from.  This is analogous to
the message the can be appended by "git cherry-pick", but in this case
we make it the default because the commit that was rebased will
necessarily remain in the history.