Auto-commit staged changes in commands 'continue' and 'record'.
If there are staged changes when the user runs 'continue' or 'record',
then auto-commit them before proceeding.  This should make the
workflow significantly smoother and more consistent with "git rebase
--interactive".  (In fact, some of the logic included in this change
was derived from the implementation of "git rebase --interactive".)

For simplicity and consistency with "git rebase', also tighten up the
condition for incorporating a user merge: the scratch reference not
only must exist but it must be checked out.