When simplifying to rebase, relax the ff requirement.
When simplifying an incremental merge of BRANCH onto MASTER into a
rebase, a simplification with --branch=BRANCH would fail because
updating BRANCH to the result of the rebase is not a fast-forward
update.

But when doing a rebase, the user typically *wants* to discard the old
versions of the commits on BRANCH (just like with "git rebase").  So
allow such a simplification even if the reference update is not a
fast-forward, *provided* that the apex of the incremental merge (i.e.,
before the simplification) is a descendant of the commit pointed to by
the old value of the reference being pointed at the result.