MergeFrontier.compute_by_bisection(): Avoid redundant test merge.
If i1 is in the last column, then there is no reason to do a test
merge in the last column *again*.

This was causing an actual failure for a user when the first test
merge succeeded but the second one failed.  This change helps prevent
that error by using the cowardly way of always getting consistent
answers: don't ask the same question more than once.