Add functionality to detect build 'conflicts' while merging
This commit adds functionality to test for build failures during the merge
process. The build script is specified via the command line flag:
--build-command=make_script.

The make_script should return 0 if the source is good, 1-127 if the source
is bad, except for code 125 if the source at this commit cannot be built or
cannot be tested.

Please see the man page for git-bisect for examples of such a build script.