Store the test command in git config
This commit uses GitConfigStore to read and write the test command
using git config. The test command is stored in a different place
for each imerge. The key is of the form: imerge.name.testcommand.
The test command is removed when the imerge is finished.

Also fixed the test-build script to use the make_script in the current dir.