Allow the default for `--edit`/`--no-edit` to be configured
The commands `git imerge continue` and `git imerge record` allow an
`--edit`/`--no-edit` option to choose whether to ask the user to edit
the commit messages if a merge commit is ready to be committed. These
messages are pretty useless, so add a configuration option
`imerge.editmergemessages` that can be used to set the default to
`--no-edit`.