continue: allow empty merges
The old code detected when an automatic commit was needed based on the
existence of staged changes (and no unstaged changes). This is wrong:
sometimes a merge legitimately involves accepting the HEAD version
unchanged. In this case, there are no staged changes, so the code
wouldn't recognize the need to make a commit. The bottom line was that
the `continue` command would fail with

    Commit $i-$j was not blocking the frontier

Instead, detect the need for an automatic commit based on the following
conditions:

* A "simple" (i.e., not octopus) merge is in progress
* There are no unstaged changes in the working tree

Note: it is not necessary to commit with `--allow-empty`, because the
usual prohibition against empty commits does not apply to merge commits.