linear_ancestry(): be less picky
Before 26d4f9a41f, `linear_ancestry()` didn't actually enforce the
`--first-parent` requirement as documented. After that commit, it did.

This turned out to be a regression. Some people were relying on the old
behavior, for example in the following scenario,

    o---A---B---C---D---E---F---G    master
             \           \
      o---o---V---W---X---M---Y      feature

where there have been merges from `master` to `feature`, and now it is
time to merge `feature` back into `master`. This used to work, but now
it fails because the merge base, `E`, is not a first-parent ancestor of
`Y` (because the first parent of `M` is `X`).

So the `--first-parent` requirement is unnecessarily restrictive.
Instead, accept (without requiring the `--first-parent` option) any
situation where there is a single path connecting the `commit1` and
`commit2` within the commits in

    git log --ancestry-path commit1..commit2

Moreover, if the `--first-parent` option is specified, then allow even
the situation where there are multiple paths between `commit1` and
`commit2` (such as `A..Y` in the above diagram). In this case, choose
the path that follows the "most-first" available parent at each merge
commit (i.e., in this case, return `[B, V, W, X, M, Y]`.

Fixes #107.