reparent: add a `--commit` option
Add an option `git imerge reparent --commit=COMMIT`, which allows the
parents of an arbitrary commit to be changed. Moreover, the
descendants of COMMIT are also rewritten all the way to HEAD.

This feature was suggested by Ke Ma <k.ma@alibaba-inc.com>, and the
implementation is partly derived from PR #130 submitted by Ke Ma.