Add setup.py script to build the package using setuptools
Setuptools is a fully-featured, actively-maintained, and stable library
designed to facilitate packaging Python projects. It is the standard
library used to create Python packages by the community.

The setup.py script allows building an sdist package and a wheel
package. Installing the package with setuptools installs a gitimerge.py
module and a git-imerge command line script.

A tox configuration was added to automate testing in a virtualenv.

The README has been updated with new installation and testing instructions.